+++ CHANGELOG.md
@@ -2,8 +2,12 @@
 
 ## [Unreleased](https://github.com/today2098/go-algorithm/compare/v0.0.1...HEAD)
 
+- Create a new structure, `BinaryHeap` about priority queue [#5](https://github.com/today2098/go-algorithm/pull/5)
+- Create a new structure, `UnionFind` about disjoint-set data structure [#6](https://github.com/today2098/go-algorithm/pull/6)
+
+
 ## [v0.0.1](https://github.com/today2098/go-algorithm/compare/84c1c73...v0.0.1) (2024/12/31)
 
-- Create a new data structure, `Stack` about simple stack (LIFO) [#1](https://github.com/today2098/go-algorithm/pull/1)
-- Create a new data structure, `Queue` about simple queue (FIFO) [#3](https://github.com/today2098/go-algorithm/pull/3/files)
-- Create a new helper structure, `IO` about I/O [#3](https://github.com/today2098/go-algorithm/pull/3/files)
+- Create a new structure, `Stack` about simple stack (LIFO) [#1](https://github.com/today2098/go-algorithm/pull/1)
+- Create a new structure, `Queue` about simple queue (FIFO) [#3](https://github.com/today2098/go-algorithm/pull/3)
+- Create a new helper structure, `IO` about I/O [#3](https://github.com/today2098/go-algorithm/pull/3)